Consolidate provider driver factories
