+++ gestaltd/services/providerdrivers/authentication.go
@@ -1,4 +1,4 @@
-package provider
+package providerdrivers
 
 import (
 	"context"
@@ -7,9 +7,9 @@ import (
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/drivers/componentprovider"
 	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
 	authenticationservice "github.com/valon-technologies/gestalt/server/services/authentication"
+	"github.com/valon-technologies/gestalt/server/services/providerdrivers/componentprovider"
 	"gopkg.in/yaml.v3"
 )
 
@@ -18,7 +18,7 @@ type yamlConfig struct {
 	CallbackURL                  string `yaml:"callbackUrl"`
 }
 
-var Factory bootstrap.AuthFactory = func(node yaml.Node, deps bootstrap.Deps) (core.AuthenticationProvider, error) {
+var AuthenticationFactory bootstrap.AuthFactory = func(node yaml.Node, deps bootstrap.Deps) (core.AuthenticationProvider, error) {
 	var cfg yamlConfig
 	if err := node.Decode(&cfg); err != nil {
 		return nil, fmt.Errorf("authentication provider: parsing config: %w", err)
