+++ gestaltd/services/providerdrivers/indexeddb.go
@@ -1,0 +1,41 @@
+package providerdrivers
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/valon-technologies/gestalt/server/core/indexeddb"
+	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
+	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
+	indexeddbservice "github.com/valon-technologies/gestalt/server/services/indexeddb"
+	"github.com/valon-technologies/gestalt/server/services/providerdrivers/componentprovider"
+	"gopkg.in/yaml.v3"
+)
+
+var IndexedDBFactory bootstrap.IndexedDBFactory = func(node yaml.Node) (indexeddb.IndexedDB, error) {
+	var cfg componentprovider.YAMLConfig
+	if err := node.Decode(&cfg); err != nil {
+		return nil, fmt.Errorf("datastore provider: parsing config: %w", err)
+	}
+	prepared, err := componentprovider.PrepareExecution(componentprovider.PrepareParams{
+		Kind:                 providermanifestv1.KindIndexedDB,
+		Subject:              "datastore provider",
+		SourceMissingMessage: "no Go, Rust, or Python datastore provider source package found",
+		Config:               cfg,
+	})
+	if err != nil {
+		return nil, err
+	}
+	cfg = prepared.YAMLConfig
+
+	return indexeddbservice.NewExecutable(context.Background(), indexeddbservice.ExecConfig{
+		Command:    cfg.Command,
+		Args:       cfg.Args,
+		Env:        cfg.Env,
+		Config:     cfg.Config,
+		Egress:     cfg.EgressPolicy(""),
+		HostBinary: cfg.HostBinary,
+		Cleanup:    prepared.Cleanup,
+		Name:       cfg.Name,
+	})
+}
