+++ gestaltd/services/providerdrivers/s3.go
@@ -1,26 +1,26 @@
-package provider
+package providerdrivers
 
 import (
 	"context"
 	"fmt"
 
-	"github.com/valon-technologies/gestalt/server/core/indexeddb"
+	s3store "github.com/valon-technologies/gestalt/server/core/s3"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
-	"github.com/valon-technologies/gestalt/server/internal/drivers/componentprovider"
 	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
-	indexeddbservice "github.com/valon-technologies/gestalt/server/services/indexeddb"
+	"github.com/valon-technologies/gestalt/server/services/providerdrivers/componentprovider"
+	s3service "github.com/valon-technologies/gestalt/server/services/s3"
 	"gopkg.in/yaml.v3"
 )
 
-var Factory bootstrap.IndexedDBFactory = func(node yaml.Node) (indexeddb.IndexedDB, error) {
+var S3Factory bootstrap.S3Factory = func(node yaml.Node) (s3store.Client, error) {
 	var cfg componentprovider.YAMLConfig
 	if err := node.Decode(&cfg); err != nil {
-		return nil, fmt.Errorf("datastore provider: parsing config: %w", err)
+		return nil, fmt.Errorf("s3 provider: parsing config: %w", err)
 	}
 	prepared, err := componentprovider.PrepareExecution(componentprovider.PrepareParams{
-		Kind:                 providermanifestv1.KindIndexedDB,
-		Subject:              "datastore provider",
-		SourceMissingMessage: "no Go, Rust, or Python datastore provider source package found",
+		Kind:                 providermanifestv1.KindS3,
+		Subject:              "s3 provider",
+		SourceMissingMessage: "no Go, Rust, or Python s3 provider source package found",
 		Config:               cfg,
 	})
 	if err != nil {
@@ -28,7 +28,7 @@ var Factory bootstrap.IndexedDBFactory = func(node yaml.Node) (indexeddb.Indexed
 	}
 	cfg = prepared.YAMLConfig
 
-	return indexeddbservice.NewExecutable(context.Background(), indexeddbservice.ExecConfig{
+	return s3service.NewExecutable(context.Background(), s3service.ExecConfig{
 		Command:    cfg.Command,
 		Args:       cfg.Args,
 		Env:        cfg.Env,
