+++ gestaltd/services/providerdrivers/workflow.go
@@ -1,27 +1,27 @@
-package provider
+package providerdrivers
 
 import (
 	"context"
 	"fmt"
 
-	coreagent "github.com/valon-technologies/gestalt/server/core/agent"
+	coreworkflow "github.com/valon-technologies/gestalt/server/core/workflow"
 	"github.com/valon-technologies/gestalt/server/internal/bootstrap"
-	"github.com/valon-technologies/gestalt/server/internal/drivers/componentprovider"
 	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
-	agentservice "github.com/valon-technologies/gestalt/server/services/agents"
+	"github.com/valon-technologies/gestalt/server/services/providerdrivers/componentprovider"
 	"github.com/valon-technologies/gestalt/server/services/runtimehost"
+	workflowservice "github.com/valon-technologies/gestalt/server/services/workflows"
 	"gopkg.in/yaml.v3"
 )
 
-var Factory bootstrap.AgentFactory = func(ctx context.Context, name string, node yaml.Node, hostServices []runtimehost.HostService, deps bootstrap.Deps) (coreagent.Provider, error) {
+var WorkflowFactory bootstrap.WorkflowFactory = func(ctx context.Context, name string, node yaml.Node, hostServices []runtimehost.HostService, deps bootstrap.Deps) (coreworkflow.Provider, error) {
 	var cfg componentprovider.YAMLConfig
 	if err := node.Decode(&cfg); err != nil {
-		return nil, fmt.Errorf("agent provider: parsing config: %w", err)
+		return nil, fmt.Errorf("workflow provider: parsing config: %w", err)
 	}
 	prepared, err := componentprovider.PrepareExecution(componentprovider.PrepareParams{
-		Kind:                 providermanifestv1.KindAgent,
-		Subject:              "agent provider",
-		SourceMissingMessage: "no Go, Rust, Python, or TypeScript agent provider source package found",
+		Kind:                 providermanifestv1.KindWorkflow,
+		Subject:              "workflow provider",
+		SourceMissingMessage: "no Go, Rust, or Python workflow provider source package found",
 		Config:               cfg,
 	})
 	if err != nil {
@@ -29,7 +29,7 @@ var Factory bootstrap.AgentFactory = func(ctx context.Context, name string, node
 	}
 	cfg = prepared.YAMLConfig
 
-	return agentservice.NewExecutable(ctx, agentservice.ExecConfig{
+	return workflowservice.NewExecutable(ctx, workflowservice.ExecConfig{
 		Command:      cfg.Command,
 		Args:         cfg.Args,
 		Env:          cfg.Env,
@@ -39,6 +39,5 @@ var Factory bootstrap.AgentFactory = func(ctx context.Context, name string, node
 		Cleanup:      prepared.Cleanup,
 		HostServices: hostServices,
 		Name:         name,
-		Telemetry:    deps.Telemetry,
 	})
 }
